Good stopping point
Still lots of compile errors but I'm getting there.
